config/requirements.txt
asgiref==3.5.2
Django==4.1.3
Pillow==9.3.0
psycopg2==2.9.5
sqlparse==0.4.3

djangorestframework~=3.14.0
django-filter~=22.1
djoser~=2.1.0